refactor: (*) Rename service.ts -> utils.ts
