+++ src/main.ts
@@ -1,5 +1,5 @@
 import { SweetId, SweetIdSize } from "./const.ts";
-import { generate, getIdLength, validate } from "./service.ts";
+import { generate, getIdLength, validate } from "./utils.ts";
 
 export function sweetid(size: SweetIdSize = "short"): SweetId {
   const id = generate(getIdLength(size));
